+++ oeps_backend/helpers/credentials.py
@@ -1,33 +1,44 @@
 import os
-from google.auth import impersonated_credentials
+import dotenv
 from google.auth.transport.requests import Request
 from google.oauth2 import service_account
 from google.cloud.bigquery import (
     Client,
-    Table,
-    SchemaField
 )
-
-import dotenv
 
 dotenv.load_dotenv()
 
-LOCAL_DATA_DIR = os.path.join(os.path.dirname(__file__), 'data_local')
-TABLE_DEF_DIR = os.path.join(os.path.dirname(__file__), 'table_definitions')
-
-
 def get_client():
+    """ Creates a BigQuery Client object and returns it, acquires credentials
+    through get_credentials(). """
 
     project_id = os.getenv("BQ_PROJECT_ID")
-
     credentials = get_credentials()
-
-    #create big query client
     client = Client(project=project_id, credentials=credentials)
-
     return client
 
 def get_credentials():
+    """ Creates and returns a credentials object to be used in a BigQuery Client.
+
+    If BQ_CREDENTIALS_FILE_PATH is present, that path is used to create the creds.
+
+    If not, the following suite of environment variables must all be present and
+    they will be used in place of the file:
+
+        BQ_PROJECT_ID
+        BQ_CREDENTIALS_PRIVATE_KEY_ID
+        BQ_CREDENTIALS_PRIVATE_KEY
+        BQ_CREDENTIALS_CLIENT_EMAIL
+        BQ_CREDENTIALS_CLIENT_ID
+        BQ_CREDENTIALS_AUTH_URI
+        BQ_CREDENTIALS_TOKEN_URI
+        BQ_CREDENTIALS_AUTH_PROVIDER_X509_CERT_URL
+        BQ_CREDENTIALS_CLIENT_X509_CERT_URL
+        BQ_CREDENTIALS_UNIVERSE_DOMAIN
+
+    These variables are all acquired through the get_credentials_info() helper function."""
+
+    credentials = None
 
     # prefer a local path to JSON credentials file.
     json_crendentials_path = os.getenv('BQ_CREDENTIALS_FILE_PATH')
@@ -36,8 +47,6 @@ def get_credentials():
             json_crendentials_path,
             scopes=["https://www.googleapis.com/auth/cloud-platform"],
         )
-        if credentials.expired and credentials.refresh_token:
-            credentials.refresh(Request())
 
     # if no local file, attempt to get all credential info from environment
     else:
@@ -46,12 +55,15 @@ def get_credentials():
             json_crendentials,
             scopes=["https://www.googleapis.com/auth/cloud-platform"],
         )
-        if credentials.expired and credentials.refresh_token:
-            credentials.refresh(Request())
+
+    if credentials and credentials.expired and credentials.refresh_token:
+        credentials.refresh(Request())
 
     return credentials
 
 def get_credentials_info():
+    """ Helper function to aggregate all BiqQuery credential info stored in environment
+    variables and return them in single dict. """
 
     return {
         "type": "service_account",
